refactor: move preview-email-template-styles to lib directory
Move non-component .ts file from components/ to lib/ where it belongs.
